backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openpyxl==3.1.5
reportlab==4.0.9
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
groq==0.4.1
websockets==12.0
# Database drivers
psycopg2-binary==2.9.11
sqlalchemy==2.0.23
# RAG dependencies
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.4
# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
